+++ src/types/index.js
@@ -1,5 +1,6 @@
 export { default as Node } from './Node'
 export { default as Magna } from './Magna'
+export { default as ExtendablePlugin } from './ExtendablePlugin'
 export { default as Import } from './Import'
 export { default as Lazy } from './Lazy'
 export { default as Module } from './Module'